pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6,<0.9.0",
    "openai>=2.30.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
